requirement.txt
tensorflow-gpu==1.11
kaldi_io
editdistance